% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_class_model.R
\name{class_model_predict}
\alias{class_model_predict}
\title{Predict the class abundance}
\usage{
class_model_predict(x, model_c = NULL, model_r)
}
\arguments{
\item{x}{a matrix of data frame. These are the predictor variables
(i.e. domain abundances).}

\item{model_c}{a model object. This is the binary classification model
returned by class_model_train(). The predictor variables used to train
the model should be exactly the same as the variables in x.}

\item{model_r}{a model object. This is the regression model returned by
class_model_train(). The predictor variables used to train
the model should be exactly the same as the variables in x.}
}
\value{
A numeric vector containing the predicted abundances.
}
\description{
Predict the class abundance
}
\examples{
class_model_predict(
model_c = t1pks_binary_model,
model_r = t1pks_regression_model)
}
