% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_class_model.R
\name{class_model_train}
\alias{class_model_train}
\title{Generate the BGC class models.}
\usage{
class_model_train(y, x, binary_method, regression_method, seed = 111)
}
\arguments{
\item{y}{a numeric vector. This is the response variables
(i.e. class abundances).}

\item{x}{a matrix of data frame. These are the predictor variables
(i.e. domain abundances).}

\item{binary_method}{method used in the binary classification. Complete
match to "rf" or "svm" (random forest and support vector machine,
respectively).}

\item{regression_method}{method used in the regression. Complete match to
"rf", "svm" or "lm" (random forest and support vector machine, linear model).}

\item{seed}{a number. Seed used to compute the random forest and support
vector machine, if these are selected as binary or regression methods.}
}
\value{
A list containting the call and the binary and regression models.
}
\description{
Generate the BGC class models.
}
\examples{
class_model_train(
y = t1pks_class_abund,
x = dom_abund,
binary_method = "rf",
regression_method = "lm",
seed = 123)
}
