% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_class_model.R
\name{get_domains}
\alias{get_domains}
\title{Map BGC class to domains}
\usage{
get_domains(bgc_class)
}
\arguments{
\item{bgc_class}{a character. Name of the target BGC class.}

\item{class2dom}{a matrix or data defining the mapping between classes and domains}
}
\value{
A character vector with domain names.
}
\description{
Map BGC class to domains
}
\examples{
get_domains("t1pks")
get_domains("nrps")
}
