% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OMs_domain}
\alias{OMs_domain}
\title{OMs domain dataset}
\format{A data frame 150 rows (metagenomes) and 82 variables (BGC domains):
\describe{
  \item{bgc_domain}{raw abundance}
  ...
}}
\usage{
OMs_domain
}
\description{
OMs dataset contains the BGC domain annotation of the synthetic marine
metagenomes created based on complete (or nearly complete) genome sequences
of the Ocean Microbial Reference Gene Catalogue (OM-RGC)
(Sunagawa et al., 2015) and the Metagenome Assembled Genomes of TARA Oceans
(MAGs) (Delmont, et al., 2017).
}
\keyword{datasets}
