% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TGs_class}
\alias{TGs_class}
\title{TGs class dataset}
\format{A data frame 150 rows (metagenomes) and 31 variables (BGC classes):
\describe{
  \item{bgc_class}{raw abundance}
  ...
}}
\usage{
TGs_class
}
\description{
TGs dataset contains the BGC class annotation of the synthetic marine
metagenomes created based on representative complete genome
sequences of the genus found in the OTU taxonomic classification of the TARA
Oceans Project (Karsenti et al., 2011).
}
\keyword{datasets}
