% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TGs_domain}
\alias{TGs_domain}
\title{TGs domain dataset}
\format{A data frame 150 rows (metagenomes) and 87 variables (BGC domains):
\describe{
  \item{bgc_domain}{raw abundance}
  ...
}}
\usage{
TGs_domain
}
\description{
TGs dataset contains the BGC domain annotation of the synthetic marine
metagenomes created based on representative complete genome
sequences of the genus found in the OTU taxonomic classification of the TARA
Oceans Project (Karsenti et al., 2011).
}
\keyword{datasets}
