% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_class_model.R
\name{wrap_up_predict}
\alias{wrap_up_predict}
\title{Predict the class abundance wrap up script}
\usage{
wrap_up_predict(x)
}
\arguments{
\item{x}{a matrix of data frame. These are the predictor variables
(i.e. domain abundances).}
}
\value{
A dataframe containing the predicted abundances.
}
\description{
Predict the class abundance wrap up script
}
\examples{
wrap_up_predict(x)
}
