% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_class_model.R
\name{wrap_up_predict}
\alias{wrap_up_predict}
\title{Predict the class abundance wrap up script}
\usage{
wrap_up_predict(x, m = models_list_oms)
}
\arguments{
\item{x}{a matrix of data frame. These are the predictor variables
(i.e. domain abundances).}

\item{m}{a list with all the models, as generated by class_model_train.}
}
\value{
A dataframe containing the predicted abundances.
}
\description{
Predict the class abundance wrap up script
}
\examples{
wrap_up_predict(x)
}
